Isolate birefringent volume static methods
